Updated compile lines due to change in gcc
Newer versions of gcc will flag changes in variable types as
critical errors rather than warnings.  This is an issue with
NetCDF b/c this library requires you to specify a variable with
the correct data type in many functions. If you are working with
different data types in the same code and use the same function
with different data types, this causes the error. The solution is to
include `-fallow-argument-mismatch` in the compile line.